Refactor SdsCertificateConfig to common package
