+++ security/pkg/nodeagent/util/util.go
@@ -20,7 +20,6 @@ import (
 	"fmt"
 	"os"
 	"path"
-	"strings"
 	"time"
 
 	"go.opencensus.io/stats/view"
@@ -105,63 +104,4 @@ func OutputKeyCertToDir(dir string, privateKey, certChain, rootCert []byte) erro
 	}
 
 	return nil
-}
-
-type SdsCertificateConfig struct {
-	CertificatePath   string
-	PrivateKeyPath    string
-	CaCertificatePath string
-}
-
-const (
-	ResourceSeparator = "~"
-)
-
-// GetResourceName converts a SdsCertificateConfig to a string to be used as an SDS resource name
-func (s SdsCertificateConfig) GetResourceName() string {
-	if s.IsKeyCertificate() {
-		return "file-cert:" + s.CertificatePath + ResourceSeparator + s.PrivateKeyPath // Format: file-cert:%s~%s
-	}
-	return ""
-}
-
-// GetRootResourceName converts a SdsCertificateConfig to a string to be used as an SDS resource name for the root
-func (s SdsCertificateConfig) GetRootResourceName() string {
-	if s.IsRootCertificate() {
-		return "file-root:" + s.CaCertificatePath // Format: file-root:%s
-	}
-	return ""
-}
-
-// IsRootCertificate returns true if this config represents a root certificate config.
-func (s SdsCertificateConfig) IsRootCertificate() bool {
-	return s.CaCertificatePath != ""
-}
-
-// IsKeyCertificate returns true if this config represents key certificate config.
-func (s SdsCertificateConfig) IsKeyCertificate() bool {
-	return s.CertificatePath != "" && s.PrivateKeyPath != ""
-}
-
-// SdsCertificateConfigFromResourceName converts the provided resource name into a SdsCertificateConfig
-// If the resource name is not valid, false is returned.
-func SdsCertificateConfigFromResourceName(resource string) (SdsCertificateConfig, bool) {
-	if strings.HasPrefix(resource, "file-cert:") {
-		filesString := strings.TrimPrefix(resource, "file-cert:")
-		split := strings.Split(filesString, ResourceSeparator)
-		if len(split) != 2 {
-			return SdsCertificateConfig{}, false
-		}
-		return SdsCertificateConfig{split[0], split[1], ""}, true
-	} else if strings.HasPrefix(resource, "file-root:") {
-		filesString := strings.TrimPrefix(resource, "file-root:")
-		split := strings.Split(filesString, ResourceSeparator)
-
-		if len(split) != 1 {
-			return SdsCertificateConfig{}, false
-		}
-		return SdsCertificateConfig{"", "", split[0]}, true
-	} else {
-		return SdsCertificateConfig{}, false
-	}
 }
